Set Client ID UI ,and render client order Details, Fix issue ko Biometric,
