+++ src/styles/global.css
@@ -33,4 +33,3 @@ body {
   height: 100%;
 }
 
-
